Move prepCrossfilterData() from DataContext to utils
